Added rotary encoder support to mugen and mugen_prototype.
