+++ keyboards/anjheos/mugen/info.json
@@ -40,7 +40,7 @@
                 { "label": "F11", "x": 11.75, "y": 0 },
                 { "label": "F12", "x": 12.75, "y": 0 },
                 { "label": "PrtSc", "x": 14, "y": 0 },
-                { "label": "Delete", "x": 15.25, "y": 0.5 },
+                { "label": "Delete", "x": 15.25, "y": 0 },
                 
 
                 { "label": "~", "x": 0, "y": 1.25 },
