--- keyboards/anjheos/mugen/keymaps/via/readme.md
@@ -1,1 +1,0 @@
-# The VIA keymap for the Mugen 75 Prototype
