--- keyboards/anjheos/mugen/keymaps/via/rules.mk
@@ -1,2 +1,0 @@
-VIA_ENABLE = yes
-
