+++ keyboards/anjheos/mugen/keymaps/vial/config.h
@@ -5,3 +5,7 @@
 #pragma once
 
 #define VIAL_KEYBOARD_UID {0x8C, 0x4F, 0xA7, 0x76, 0xE4, 0x4D, 0xCF, 0x78}
+
+#define VIAL_ENCODER_DEFAULT {KC_VOLD, KC_VOLU, KC_WH_D, KC_WH_U, KC_WH_L, KC_WH_R, KC_TRNS, KC_TRNS,}
+
+#define VIAL_ENCODER_KEYCODE_DELAY 10
