+++ keyboards/anjheos/mugen/keymaps/vial/keymap.c
@@ -52,14 +52,3 @@ const uint16_t PROGMEM keymaps[][MATRIX_ROWS][MATRIX_COLS] = {
 		
 };
 
-bool encoder_update_user(uint8_t index, bool clockwise) {
-	if (index == 0){
-		if (clockwise) {
-			tap_code(KC_1);
-		} else {
-			tap_code(KC_2);
-		}
-	}
-	
-	return true;
-}
