+++ keyboards/anjheos/mugen/mugen.h
@@ -20,14 +20,6 @@
 
 #define XXX KC_NO
 
-/* This is a shortcut to help you visually see your layout.
- *
- * The first section contains all of the arguments representing the physical
- * layout of the board and position of the keys.
- *
- * The second converts the arguments into a two-dimensional array which
- * represents the switch matrix.
- */
 #define LAYOUT_all( \
 	k00,    k01, k02, k03, k04,    k05, k06, k07, k08,     k09, k0a, k0b, k0c, 	k0d, k0e, \
 	k10, k11, k12, k13, k14, k15, k16, k17, k18, k19, k1a, k1b, k1c, k3d, k1d, 		 k1e, \
