+++ keyboards/anjheos/mugen/rules.mk
@@ -1,1 +1,2 @@
-# This file intentionally left blank
+# Copyright 2022 Anjheos (@Anjheos)
+ENCODER_ENABLE = yes
