+++ keyboards/anjheos/mugen_prototype/keymaps/via/keymap.c
@@ -53,13 +53,19 @@ const uint16_t PROGMEM keymaps[][MATRIX_ROWS][MATRIX_COLS] = {
 };
 
 bool encoder_update_user(uint8_t index, bool clockwise) {
-	if (index == 0){
+	if (get_highest_layer(layer_state|default_layer_state) == 0) {
 		if (clockwise) {
-			tap_code(KC_1);
+			tap_code_delay(KC_VOLU, 10);
 		} else {
-			tap_code(KC_2);
+			tap_code_delay(KC_VOLD, 10);
+		}
+	} else {
+		if (clockwise) {
+			tap_code(KC_WH_U);
+		} else {
+			tap_code(KC_WH_D);
 		}
 	}
-	
-	return true;
+	return false;
 }
+
