+++ keyboards/anjheos/mugen_prototype/keymaps/vial/config.h
@@ -5,3 +5,7 @@
 #pragma once
 
 #define VIAL_KEYBOARD_UID {0x31, 0xC3, 0x43, 0x59, 0x49, 0x66, 0x0C, 0x8C}
+
+#define VIAL_ENCODER_DEFAULT {KC_VOLD, KC_VOLU, KC_WH_D, KC_WH_U, KC_WH_L, KC_WH_R, KC_TRNS, KC_TRNS,}
+
+#define VIAL_ENCODER_KEYCODE_DELAY 10
