+++ keyboards/anjheos/mugen_prototype/keymaps/vial/keymap.c
@@ -51,15 +51,3 @@ const uint16_t PROGMEM keymaps[][MATRIX_ROWS][MATRIX_COLS] = {
 	_______, _______, _______, 							  _______, 							  _______, _______, _______, _______, _______)
 		
 };
-
-bool encoder_update_user(uint8_t index, bool clockwise) {
-	if (index == 0){
-		if (clockwise) {
-			tap_code(KC_1);
-		} else {
-			tap_code(KC_2);
-		}
-	}
-	
-	return true;
-}
