+++ keyboards/anjheos/mugen_prototype/rules.mk
@@ -1,1 +1,2 @@
 # This file intentionally left blank
+ENCODER_ENABLE = yes
